Moved client tests into tests; Moved client main() into examples/cli-client.rs
